Refactor image processing in to a module
