+++ webify_models/src/image_processing/image.rs
@@ -1,0 +1,7 @@
+use std::path::PathBuf;
+
+#[derive(Debug, Clone)]
+pub struct Image {
+    pub path: PathBuf,
+    pub extension: String,
+}
