+++ webify_models/src/image_processing/mod.rs
@@ -1,0 +1,12 @@
+pub mod convert_to_png;
+pub mod scan_dir_for_images;
+pub mod move_to_textures_dir;
+pub mod image;
+pub mod process;
+
+pub use self::image::Image;
+
+pub use self::convert_to_png::convert_to_png;
+pub use self::scan_dir_for_images::scan_dir_for_images;
+pub use self::move_to_textures_dir::move_to_textures_dir;
+pub use self::process::process;
