+++ webify_models/src/mesh_update/mod.rs
@@ -1,0 +1,3 @@
+mod scan_dir_for_meshes;
+
+pub use self::scan_dir_for_meshes::scan_dir_for_meshes;
